requirements.txt
requests>=2.28.0
pyyaml>=6.0
psutil>=5.9.0
apscheduler>=3.10.0

# Sensor libraries (install only what you need)
# 1-Wire sensors
w1thermsensor

# I2C sensors
smbus

# GPIO sensors
gpiozero

# Adafruit sensors
adafruit-circuitpython-htu21d
adafruit-circuitpython-tsl2561
Adafruit-BMP
adafruit-circuitpython-adxl34x
adafruit-circuitpython-dht
adafruit-circuitpython-vl53l0x

# Network tools
pingparsing

# Utilities
deepdiff

# Modbus (for SDM120)
sdm_modbus

# RPi specific (optional)
# git+https://github.com/nicmcd/vcgencmd.git
